Fixed comment,deleted Create Reservation.java to move in the separated PR, reverted changes in the HypedisksIT
