+++ compute/cloud-client/src/test/java/compute/disks/HypedisksIT.java
@@ -14,32 +14,37 @@
  * limitations under the License.
  */
 
-package compute.disks.storagepool;
+package compute.disks;
 
 import static com.google.common.truth.Truth.assertWithMessage;
 
 import com.google.cloud.compute.v1.Disk;
 import com.google.cloud.compute.v1.StoragePool;
 import com.google.cloud.compute.v1.StoragePoolsClient;
-import compute.disks.DeleteDisk;
+import compute.disks.storagepool.CreateDiskInStoragePool;
+import compute.disks.storagepool.CreateHyperdiskStoragePool;
 import java.io.IOException;
 import java.util.UUID;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 import org.junit.Assert;
+import org.junit.FixMethodOrder;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.Timeout;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
+import org.junit.runners.MethodSorters;
 
 @RunWith(JUnit4.class)
 @Timeout(value = 40, unit = TimeUnit.MINUTES)
-class CreateDiskInStoragePoolTest {
+@FixMethodOrder(MethodSorters.NAME_ASCENDING)
+public class HypedisksIT {
   private static final String PROJECT_ID = System.getenv("GOOGLE_CLOUD_PROJECT");
-  private static final String ZONE = "me-central1-b";
+  private static final String ZONE = "europe-west1-b";
+  private static String HYPERDISK_NAME;
   private static String HYPERDISK_IN_POOL_NAME;
   private static String STORAGE_POOL_NAME;
 
@@ -50,33 +55,65 @@ class CreateDiskInStoragePoolTest {
   }
 
   @BeforeAll
-  public static void setUp()
-      throws IOException, ExecutionException, InterruptedException, TimeoutException {
+  public static void setUp() {
     requireEnvVar("GOOGLE_APPLICATION_CREDENTIALS");
     requireEnvVar("GOOGLE_CLOUD_PROJECT");
 
+    HYPERDISK_NAME = "test-hyperdisk-enc-" + UUID.randomUUID();
     HYPERDISK_IN_POOL_NAME = "test-hyperdisk-enc-" + UUID.randomUUID();
     STORAGE_POOL_NAME = "test-storage-pool-enc-" + UUID.randomUUID();
-    String poolType = String.format("projects/%s/zones/%s/storagePoolTypes/hyperdisk-balanced",
-        PROJECT_ID, ZONE);
-    StoragePool storagePool = CreateHyperdiskStoragePool
-        .createHyperdiskStoragePool(PROJECT_ID, ZONE, STORAGE_POOL_NAME, poolType,
-            "advanced", 10240, 10000, 10240);
   }
 
   @AfterAll
   public static void cleanup()
       throws IOException, InterruptedException, ExecutionException, TimeoutException {
     // Delete all disks created for testing.
+    DeleteDisk.deleteDisk(PROJECT_ID, ZONE, HYPERDISK_NAME);
     DeleteDisk.deleteDisk(PROJECT_ID, ZONE, HYPERDISK_IN_POOL_NAME);
-    // Delete Storage Pool created for testing.
+
     try (StoragePoolsClient client = StoragePoolsClient.create()) {
       client.deleteAsync(PROJECT_ID, ZONE, STORAGE_POOL_NAME);
     }
   }
 
   @Test
-  public void createHyperdiskInStoragePoolTest()
+  public void stage1_CreateHyperdiskTest()
+      throws IOException, ExecutionException, InterruptedException, TimeoutException {
+    String diskType = String.format("zones/%s/diskTypes/hyperdisk-balanced", ZONE);
+
+    Disk hyperdisk = CreateHyperdisk
+        .createHyperdisk(PROJECT_ID, ZONE, HYPERDISK_NAME, diskType, 10, 3000, 140);
+
+    Assert.assertNotNull(hyperdisk);
+    Assert.assertEquals(HYPERDISK_NAME, hyperdisk.getName());
+    Assert.assertEquals(3000, hyperdisk.getProvisionedIops());
+    Assert.assertEquals(140, hyperdisk.getProvisionedThroughput());
+    Assert.assertEquals(10, hyperdisk.getSizeGb());
+    Assert.assertTrue(hyperdisk.getType().contains("hyperdisk-balanced"));
+    Assert.assertTrue(hyperdisk.getZone().contains(ZONE));
+  }
+
+  @Test
+  public void stage1_CreateHyperdiskStoragePoolTest()
+      throws IOException, ExecutionException, InterruptedException, TimeoutException {
+    String poolType = String.format("projects/%s/zones/%s/storagePoolTypes/hyperdisk-balanced",
+        PROJECT_ID, ZONE);
+    StoragePool storagePool = CreateHyperdiskStoragePool
+        .createHyperdiskStoragePool(PROJECT_ID, ZONE, STORAGE_POOL_NAME, poolType,
+            "advanced", 10240, 20000, 10240);
+
+    Assert.assertNotNull(storagePool);
+    Assert.assertEquals(STORAGE_POOL_NAME, storagePool.getName());
+    Assert.assertEquals(20000, storagePool.getPoolProvisionedIops());
+    Assert.assertEquals(10240, storagePool.getPoolProvisionedThroughput());
+    Assert.assertEquals(10240, storagePool.getPoolProvisionedCapacityGb());
+    Assert.assertTrue(storagePool.getStoragePoolType().contains("hyperdisk-balanced"));
+    Assert.assertTrue(storagePool.getCapacityProvisioningType().equalsIgnoreCase("advanced"));
+    Assert.assertTrue(storagePool.getZone().contains(ZONE));
+  }
+
+  @Test
+  public void stage2_CreateHyperdiskStoragePoolTest()
       throws IOException, ExecutionException, InterruptedException, TimeoutException {
     String diskType = String.format("zones/%s/diskTypes/hyperdisk-balanced", ZONE);
     String storagePoolLink = String
